fixup! src: prefer v8::Global over node::Persistent
